Remove conflicting tests from core test blanks
Extract contracting-v1 specific tests
Restore Mixins for Contracting v3
