+++ openprocurement/auctions/core/tests/chronograph.py
@@ -1,0 +1,15 @@
+from openprocurement.auctions.core.tests.base import snitch
+from openprocurement.auctions.core.tests.blanks.chronograph_blanks import (
+    contract_signing_period_switch_to_qualification,
+    contract_signing_period_switch_to_complete,
+)
+
+
+class AuctionContractSwitchTestMixin(object):
+    test_contract_signing_period_switch_to_qualification = snitch(
+        contract_signing_period_switch_to_qualification
+    )
+
+    test_contract_signing_period_switch_to_complete = snitch(
+        contract_signing_period_switch_to_complete
+    )
